rename module names to be closer to pep8 recomendations.
https://www.python.org/dev/peps/pep-0008/#package-and-module-names
